requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
email-validator==2.3.0
bcrypt==5.0.0
apscheduler==3.11.2
requests==2.32.3
python-dotenv==1.2.2